refactor(suite): extract shared debug utilities
