+++ packages/suite/src/components/firmware/FirmwareOffer.tsx
@@ -1,3 +1,4 @@
+import { DebugOnlyBadge } from '@suite/debug';
 import { getSuiteFirmwareTypeString, useFirmwareDesktopUpdate } from '@suite/firmware-upgrade';
 import { Translation, useTranslation } from '@suite/intl';
 import { selectIsDebugModeActive } from '@suite/settings';
@@ -13,8 +14,6 @@ import { spacings } from '@trezor/theme';
 
 import { MarkdownWithComponents } from 'src/components/suite';
 import { useSelector } from 'src/hooks/suite';
-
-import { DebugOnlyBadge } from '../suite/DebugOnlyBadge';
 
 type FirmwareOfferProps = {
     isCustomFirmware?: boolean;
