+++ packages/suite/src/views/wallet/send/Outputs/DevSelfAddress.tsx
@@ -1,9 +1,9 @@
+import { DebugOnlyBadge } from '@suite/debug';
 import { getFirstFreshAddress } from '@suite-common/address';
 import { type Account } from '@suite-common/wallet-types';
 import { isUtxoBased } from '@suite-common/wallet-utils';
 import { Button } from '@trezor/components';
 
-import { DebugOnlyBadge } from 'src/components/suite/DebugOnlyBadge';
 import { useSendFormContext } from 'src/hooks/wallet';
 
 type DevSelfAddressProps = {
