+++ packages/suite/src/views/wallet/staking/components/StakingDashboard/components/DebugOnlyCardanoStakingCard.tsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
 
+import { DebugOnlyBadge } from '@suite/debug';
 import { selectIsDebugModeActive } from '@suite/settings';
 import { selectCardanoPoolsInfo } from '@suite-common/wallet-core';
 import { type Account } from '@suite-common/wallet-types';
@@ -10,7 +11,6 @@ import {
 import { Card, Column, Icon, InfoItem, Paragraph, Row } from '@trezor/components';
 import { spacings } from '@trezor/theme';
 
-import { DebugOnlyBadge } from 'src/components/suite/DebugOnlyBadge';
 import { useSelector } from 'src/hooks/suite';
 
 const ParagraphWrapper = styled.div`
