+++ suite/debug/src/DebugOnlyBadge.tsx
@@ -1,9 +1,11 @@
+import { type ReactNode } from 'react';
+
 import { Translation } from '@suite/intl';
 import { Badge, Row } from '@trezor/components';
 import { spacings } from '@trezor/theme';
 
 type DebugOnlyBadgeProps = {
-    children?: React.ReactNode;
+    children?: ReactNode;
 };
 
 export const DebugOnlyBadge = ({ children }: DebugOnlyBadgeProps) => (
