+++ suite/debug/src/index.ts
@@ -1,0 +1,5 @@
+export * from './DebugLegend';
+export * from './DebugOnly';
+export * from './DebugOnlyBadge';
+export * from './useDebugLanguageShortcut';
+export * from './useDebugModeActivator';
