+++ suite/debug/tsconfig.json
@@ -1,0 +1,17 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "compilerOptions": { "outDir": "libDev" },
+    "references": [
+        {
+            "path": "../../suite-common/suite-types"
+        },
+        { "path": "../intl" },
+        { "path": "../settings" },
+        { "path": "../../packages/components" },
+        {
+            "path": "../../packages/suite-desktop-api"
+        },
+        { "path": "../../packages/theme" },
+        { "path": "../../packages/utils" }
+    ]
+}
